Uploading  benchmarks now works with the database. Need to clean up the code and refactor a bit though.
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/starexec@5632 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/data/to/Benchmark.java
@@ -7,7 +7,7 @@ public class Benchmark {
 	private long id;
 	private String path;
 	private String fileName;
-	private long userid;
+	private int userid;
 	private Date uploaded;
 	private int level;
 	
@@ -28,11 +28,11 @@ public class Benchmark {
 		fileName = path.substring(path.lastIndexOf(File.separator) + 1);
 	}
 	
-	public long getUserId() {
+	public int getUserId() {
 		return userid;
 	}
 	
-	public void setUserId(long userid) {
+	public void setUserId(int userid) {
 		this.userid = userid;
 	}
